A design-pattern fueled simple Evolutionary Algorithm
This EA tires to find the solution of a mathematical system that I invented at random. The important stuff is the design pattern organization. The EA tries to be very generic, allowing many different operators for crossover, mutation, selection and stop criteria to be implement using the Strategy Pattern. A factory method helps initializing the EA. And a singleton fixes the random for consistent tests. Also, the ScottPlot package helps us to create an image with the line plot of the evolution.